TASK: Move generic `Commands` into cr core
